Add files via upload
ChapterTest 
Q1 : Create a variable of type string and try to add a number to it.
Q2 :  Use typeof operator to find the datatype of the string in last question.
Q3 : Create a const object in javascript. Can you change it to hold a number later?
Q4 : Try to add a new to key to the const object in problem 3. Were uou able to do it?
Q5 : Write a Js program to create a word-meaning dictionary of 2 words.